pyproject.toml
[project]
name = "csd-peak-identifier"
version = "0.1.0"
description = "A PySide6 GUI for identifying isotopes in ECR ion beam spectra."
authors = [
	{name="Jessica Rehak", email="<rehak@lbl.gov>"}
]
requires-python = ">=3.13,<3.15"
dependencies = ["pyside6 (>=6.11.0,<7.0.0)", "numpy (>=2.4.4,<3.0.0)", "pandas (>=3.0.2,<4.0.0)", "matplotlib (>=3.10.8,<4.0.0)", "scipy (>=1.17.1,<2.0.0)", "ops-ecris-analysis @ git+https://github.com/lbnl-88ops/ecris.analysis.git@v1.4.3", "requests (>=2.33.1,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
csd-peak-identifier = "csd_peak_identifier.main:main"
